+++ src/components/Pages/CallbackPage/CallbackPage.tsx
@@ -11,7 +11,8 @@ import { ConnectionType } from '../../../modules/analytics/types'
 import { useCurrentConnectionData } from '../../../shared/connection'
 import { isMagicExtensionError, isMagicRpcError } from '../../../shared/errors'
 import { extractReferrerFromSearchParameters, locations } from '../../../shared/locations'
-import { getConnectionOptionFromState, isMobileSession } from '../../../shared/mobile'
+import { isMobileSession } from '../../../shared/mobile'
+import { getConnectionOptionFromState } from '../../../shared/oauthState'
 import { getStoredEmail } from '../../../shared/onboarding/getStoredEmail'
 import { markReturningUser } from '../../../shared/onboarding/markReturningUser'
 import { trackCheckpoint } from '../../../shared/onboarding/trackCheckpoint'
