+++ src/components/Pages/MobileCallbackPage/MobileCallbackPage.tsx
@@ -9,7 +9,7 @@ import { ConnectionType } from '../../../modules/analytics/types'
 import { createAuthServerHttpClient } from '../../../shared/auth'
 import { isMagicRpcError } from '../../../shared/errors'
 import { locations } from '../../../shared/locations'
-import { getConnectionOptionFromState } from '../../../shared/mobile'
+import { getConnectionOptionFromState } from '../../../shared/oauthState'
 import { handleError } from '../../../shared/utils/errorHandler'
 import { OAUTH_ACCESS_DENIED_ERROR, createMagicInstance } from '../../../shared/utils/magicSdk'
 import { ConnectionContainer, ConnectionTitle, DecentralandLogo, ProgressContainer } from '../../ConnectionModal/ConnectionLayout.styled'
