+++ src/shared/mobile.spec.ts
@@ -1,5 +1,5 @@
-import { ConnectionOptionType } from '../components/Connection'
-import { getConnectionOptionFromState, getMobileSession, isMobileSession, resetMobileSession } from './mobile'
+import { getMobileSession, isMobileSession, resetMobileSession } from './mobile'
+import { resetOauthStateCache } from './oauthState'
 
 function setLocation(pathname: string, search = '') {
   Object.defineProperty(window, 'location', {
@@ -15,6 +15,7 @@ function makeState(customData: Record<string, unknown>): string {
 describe('mobile session', () => {
   beforeEach(() => {
     resetMobileSession()
+    resetOauthStateCache()
     setLocation('/auth/login')
   })
 
@@ -216,78 +217,6 @@ describe('mobile session', () => {
         it('should return null', () => {
           expect(getMobileSession()).toBeNull()
         })
-      })
-    })
-  })
-
-  describe('getConnectionOptionFromState', () => {
-    describe('when state has a known connectionOption', () => {
-      beforeEach(() => {
-        const state = makeState({ connectionOption: ConnectionOptionType.GOOGLE })
-        setLocation('/auth/callback', `?state=${state}`)
-      })
-
-      it('should return the ConnectionOptionType value', () => {
-        expect(getConnectionOptionFromState()).toBe(ConnectionOptionType.GOOGLE)
-      })
-    })
-
-    describe('when state has an unknown connectionOption', () => {
-      beforeEach(() => {
-        const state = makeState({ connectionOption: 'not-a-real-option' })
-        setLocation('/auth/callback', `?state=${state}`)
-      })
-
-      it('should return undefined to prevent typos leaking into tracking', () => {
-        expect(getConnectionOptionFromState()).toBeUndefined()
-      })
-    })
-
-    describe('when state has no connectionOption field', () => {
-      beforeEach(() => {
-        const state = makeState({ isMobileFlow: true })
-        setLocation('/auth/callback', `?state=${state}`)
-      })
-
-      it('should return undefined', () => {
-        expect(getConnectionOptionFromState()).toBeUndefined()
-      })
-    })
-
-    describe('when there is no state param', () => {
-      beforeEach(() => {
-        setLocation('/auth/callback')
-      })
-
-      it('should return undefined', () => {
-        expect(getConnectionOptionFromState()).toBeUndefined()
-      })
-    })
-
-    describe('when state is malformed', () => {
-      beforeEach(() => {
-        setLocation('/auth/callback', '?state=not-base64!!!')
-      })
-
-      it('should return undefined', () => {
-        expect(getConnectionOptionFromState()).toBeUndefined()
-      })
-    })
-
-    describe('caching across reads', () => {
-      beforeEach(() => {
-        const state = makeState({ connectionOption: ConnectionOptionType.DISCORD })
-        setLocation('/auth/callback', `?state=${state}`)
-      })
-
-      it('should still return the option after the state param has been stripped from the URL', () => {
-        // First read populates the in-memory cache while the state param is present
-        expect(getConnectionOptionFromState()).toBe(ConnectionOptionType.DISCORD)
-
-        // Simulate Magic's getRedirectResult() wiping the query string
-        setLocation('/auth/callback', '')
-
-        expect(getConnectionOptionFromState()).toBe(ConnectionOptionType.DISCORD)
       })
     })
   })
