+++ src/shared/mobile.ts
@@ -1,55 +1,12 @@
-import { ConnectionOptionType } from '../components/Connection'
+import { extractCustomDataFromState } from './oauthState'
 
 interface MobileSession {
   u?: string
   s?: string
 }
 
-interface OAuthCustomState {
-  isMobileFlow?: boolean
-  mobileUserId?: string
-  mobileSessionId?: string
-  connectionOption?: ConnectionOptionType
-  redirectTo?: string
-  referrer?: string
-}
-
-// In-memory caches: survive SPA navigations without localStorage. Module-scoped so we don't
-// re-parse the state param on every read (and so consumers can read it after Magic's
-// getRedirectResult() strips it from the URL, as long as the first read happened earlier).
+// In-memory cache: survives SPA navigations without localStorage
 let cachedSession: MobileSession | null | undefined
-let cachedCustomData: OAuthCustomState | null | undefined
-
-// Parses the OAuth `state` query param produced by the social-login redirect.
-// Returns null on missing/malformed state.
-function extractCustomDataFromState(): OAuthCustomState | null {
-  if (cachedCustomData !== undefined) {
-    return cachedCustomData
-  }
-
-  try {
-    const params = new URLSearchParams(window.location.search)
-    const state = params.get('state')
-    if (state) {
-      const decoded = JSON.parse(atob(state))
-      cachedCustomData = JSON.parse(decoded.customData) as OAuthCustomState
-      return cachedCustomData
-    }
-  } catch {
-    // ignore malformed state
-  }
-  cachedCustomData = null
-  return null
-}
-
-function isConnectionOptionType(value: unknown): value is ConnectionOptionType {
-  return typeof value === 'string' && (Object.values(ConnectionOptionType) as string[]).includes(value)
-}
-
-function getConnectionOptionFromState(): ConnectionOptionType | undefined {
-  const value = extractCustomDataFromState()?.connectionOption
-  return isConnectionOptionType(value) ? value : undefined
-}
 
 function getMobileSession(): MobileSession | null {
   if (cachedSession !== undefined) {
@@ -84,8 +41,7 @@ function isMobileSession(): boolean {
 // Exported for testing only
 function resetMobileSession(): void {
   cachedSession = undefined
-  cachedCustomData = undefined
 }
 
-export type { MobileSession, OAuthCustomState }
-export { getMobileSession, isMobileSession, resetMobileSession, getConnectionOptionFromState }
+export type { MobileSession }
+export { getMobileSession, isMobileSession, resetMobileSession }
